Added query options
